Move to database gen package
